chartコンポーネントを整理した
