+++ common/types/chart.ts
@@ -1,0 +1,8 @@
+export type ChartDataProps = {
+  labels: string[];
+  label: string;
+  data: number[];
+  bgColor: string[];
+  bdColor: string[];
+  bdWidth: number;
+}
